+++ .gitignore
@@ -2,3 +2,4 @@
 *.json
 *.txt
 *.pyproj
+*.pyc
